Adição inicial da service

+++ MuseuTreze/src/main/java/inf/laboratorio/museutreze/repository/AssuntoRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AssuntoRepository extends JpaRepository<Assunto, Integer> {
-    Assunto findById(int id);
+public interface AssuntoRepository extends JpaRepository<Assunto, Long> {
+    Assunto findById(long id);
     Assunto findByDescricao(String descricao);
 
     List<Assunto> findByDescricaoContainingIgnoreCase(String descricao);
